rag_backend/requirements.txt
fastapi
uvicorn
qdrant-client
python-dotenv
cohere
requests
trafilatura